[RelVal] Modularise and prepare, clean
* remove some unused code

* move helper and utility functions to o2dpg_release_validation_utils.py
  and common variables to o2dpg_release_validation_variables.py in view
  of implementations that allow for more in-depth metrics and threshold
  studies

* allow to directly pass flat ROOT files ==> no further extraction run
  in this case.
  Also in view of aforementioned studies
