+++ RelVal/o2dpg_release_validation.py
@@ -74,7 +74,7 @@ import sys
 import argparse
 import re
 from os import environ, makedirs, remove
-from os.path import join, abspath, exists, isfile, isdir, dirname, relpath
+from os.path import join, abspath, exists, isdir, dirname, basename, relpath
 from glob import glob
 from subprocess import Popen, PIPE, STDOUT
 from pathlib import Path
@@ -93,275 +93,18 @@ if O2DPG_ROOT is None:
     print('ERROR: This needs O2DPG loaded')
     sys.exit(1)
 
+sys.path.append(join(dirname(__file__), '.', 'o2dpg_release_validation_variables'))
+import o2dpg_release_validation_variables as variables
+
+sys.path.append(join(dirname(__file__), '.', 'o2dpg_release_validation_utils'))
+from o2dpg_release_validation_utils import *
+
 ROOT_MACRO_EXTRACT=join(O2DPG_ROOT, "RelVal", "ExtractAndFlatten.C")
 ROOT_MACRO_RELVAL=join(O2DPG_ROOT, "RelVal", "ReleaseValidation.C")
 
 from ROOT import gROOT
 
-DETECTORS_OF_INTEREST_HITS = ["ITS", "TOF", "EMC", "TRD", "PHS", "FT0", "HMP", "MFT", "FDD", "FV0", "MCH", "MID", "CPV", "ZDC", "TPC"]
-
-REL_VAL_SEVERITIES = ["GOOD", "WARNING", "NONCRIT_NC", "CRIT_NC", "BAD"]
-REL_VAL_SEVERITIES_USE_SUMMARY = [True, False, False, True, True]
-REL_VAL_SEVERITY_MAP = {v: i for i, v in enumerate(REL_VAL_SEVERITIES)}
-REL_VAL_SEVERITY_COLOR_MAP = {"GOOD": "green", "WARNING": "orange", "NONCRIT_NC": "cornflowerblue", "CRIT_NC": "navy", "BAD": "red"}
-REL_VAL_TEST_NAMES = ["chi2", "kolmogorov", "num_entries"]
-REL_VAL_TEST_NAMES_MAP = {v: i for i, v in enumerate(REL_VAL_TEST_NAMES)}
-REL_VAL_TEST_CRITICAL = [True, True, False]
-REL_VAL_TEST_DEFAULT_THRESHOLDS = [1.5, 0.5, 0.01]
-REL_VAL_TEST_UPPER_LOWER_THRESHOLD = [1, -1, 1]
-REL_VAL_TEST_SUMMARY_NAME = "summary"
-REL_VAL_TEST_NAMES_SUMMARY = REL_VAL_TEST_NAMES + [REL_VAL_TEST_SUMMARY_NAME]
-REL_VAL_TEST_NAMES_MAP_SUMMARY = {v: i for i, v in enumerate(REL_VAL_TEST_NAMES_SUMMARY)}
-
 gROOT.SetBatch()
-
-
-def find_mutual_files(dirs, glob_pattern, *, grep=None):
-    """
-    Find mutual files recursively in list of dirs
-
-    Args:
-        dirs: iterable
-            directories to take into account
-        glob_pattern: str
-            pattern used to apply glob to only seach for some files
-        grep: iterable
-            additional list of patterns to grep for
-    Returns:
-        list: intersection of found files
-    """
-    files = []
-    for d in dirs:
-        glob_path = f"{d}/**/{glob_pattern}"
-        files.append(glob(glob_path, recursive=True))
-
-    for f, d in zip(files, dirs):
-        f.sort()
-        for i, _ in enumerate(f):
-            # strip potential leading /
-            f[i] = f[i][len(d):].lstrip("/")
-
-    # build the intersection
-    if not files:
-        return []
-
-    intersection = files[0]
-    for f in files[1:]:
-        intersection = list(set(intersection) & set(f))
-
-    # apply additional grepping if patterns are given
-    if grep:
-        intersection_cache = intersection.copy()
-        intersection = []
-        for g in grep:
-            for ic in intersection_cache:
-                if g in ic:
-                    intersection.append(ic)
-
-    # Sort for convenience
-    intersection.sort()
-
-    return intersection
-
-
-def exceeding_difference_thresh(sizes, threshold=0.1):
-    """
-    Find indices in sizes where value exceeds threshold
-    """
-    diff_indices = []
-    for i1, i2 in combinations(range(len(sizes)), 2):
-        diff = abs(sizes[i1] - sizes[i2])
-        if diff / sizes[i2] > threshold or diff / sizes[i2] > threshold:
-            diff_indices.append((i1, i2))
-    return diff_indices
-
-
-def file_sizes(dirs, threshold):
-    """
-    Compare file sizes of mutual files in given dirs
-    """
-    intersection = find_mutual_files(dirs, "*.root")
-
-    # prepare for convenient printout
-    max_col_lengths = [0] * (len(dirs) + 1)
-    sizes = [[] for _ in dirs]
-
-    # extract file sizes
-    for f in intersection:
-        max_col_lengths[0] = max(max_col_lengths[0], len(f))
-        for i, d in enumerate(dirs):
-            size = Path(join(d, f)).stat().st_size
-            max_col_lengths[i + 1] = max(max_col_lengths[i + 1], len(str(size)))
-            sizes[i].append(size)
-
-    # prepare dictionary to be dumped and prepare printout
-    collect_dict = {"directories": dirs, "files": {}, "threshold": threshold}
-    top_row = "| " + " | ".join(dirs) + " |"
-    print(f"\n{top_row}\n")
-    for i, f in enumerate(intersection):
-        compare_sizes = []
-        o = f"{f:<{max_col_lengths[0]}}"
-        for j, s in enumerate(sizes):
-            o += f" | {str(s[i]):<{max_col_lengths[j+1]}}"
-            compare_sizes.append(s[i])
-        o = f"| {o} |"
-
-        diff_indices =  exceeding_difference_thresh(compare_sizes, threshold)
-        if diff_indices:
-            o += f"  <==  EXCEEDING threshold of {threshold} at columns {diff_indices} |"
-            collect_dict["files"][f] = compare_sizes
-        else:
-            o += " OK |"
-        print(o)
-    return collect_dict
-
-
-def load_patterns(include_patterns, exclude_patterns, print_loaded=True):
-    """
-    Load include patterns to be used for regex comparion
-    """
-    def load_this_patterns(patterns):
-        if not patterns or not patterns[0].startswith("@"):
-            return patterns
-        with open(include_patterns[0][1:], "r") as f:
-            return f.read().splitlines()
-
-    include_patterns = load_this_patterns(include_patterns)
-    exclude_patterns = load_this_patterns(exclude_patterns)
-    if print_loaded:
-        if include_patterns:
-            print("Following patterns are included:")
-            for ip in include_patterns:
-                print(f"  - {ip}")
-        if exclude_patterns:
-            print("Following patterns are excluded:")
-            for ep in exclude_patterns:
-                print(f"  - {ep}")
-    return include_patterns, exclude_patterns
-
-
-def check_patterns(name, include_patterns, exclude_patterns):
-    """
-    check a name against a list of regex
-    """
-    if not include_patterns and not exclude_patterns:
-        return True
-    if include_patterns:
-        for ip in include_patterns:
-            if re.search(ip, name):
-                return True
-        return False
-    if exclude_patterns:
-        for ip in exclude_patterns:
-            if re.search(ip, name):
-                return False
-        return True
-    return False
-
-def check_flags(tests, flags, flags_summary):
-    """
-    include histograms based on the flags
-    """
-    if not flags and not flags_summary:
-        return True
-    for test in tests:
-        if test["test_name"] == REL_VAL_TEST_SUMMARY_NAME:
-            if flags_summary:
-                for f in flags_summary:
-                    if test["result"] == f:
-                        return True
-        elif flags:
-            for f in flags:
-                if test["result"] == f:
-                    return True
-    return False
-
-
-def plot_pie_charts(summary, out_dir, title, include_patterns=None, exclude_patterns=None):
-
-    print("==> Plot pie charts <==")
-
-    test_n_hist_map = {}
-
-    # need to re-arrange the JSON structure abit for per-test result pie charts
-    for histo_name, tests in summary.items():
-        # check if histo_name is in include patterns
-        if not check_patterns(histo_name, include_patterns, exclude_patterns):
-            continue
-        # loop over tests done
-        for test in tests:
-            test_name = test["test_name"]
-            if test_name not in test_n_hist_map:
-                test_n_hist_map[test_name] = {}
-            result = test["result"]
-            if result not in test_n_hist_map[test_name]:
-                test_n_hist_map[test_name][result] = 0
-            test_n_hist_map[test_name][result] += 1
-
-    for which_test, flags in test_n_hist_map.items():
-        labels = []
-        colors = []
-        n_histos = []
-        for flag, count in flags.items():
-            labels.append(flag)
-            n_histos.append(count)
-            colors.append(REL_VAL_SEVERITY_COLOR_MAP[flag])
-
-        figure, ax = plt.subplots(figsize=(20, 20))
-        ax.pie(n_histos, explode=[0.05 for _ in labels], labels=labels, autopct="%1.1f%%", startangle=90, textprops={"fontsize": 30}, colors=colors)
-        ax.axis("equal")
-        ax.axis("equal")
-
-        figure.suptitle(f"{title} ({which_test})", fontsize=40)
-        save_path = join(out_dir, f"pie_chart_{which_test}.png")
-        figure.savefig(save_path)
-        plt.close(figure)
-
-
-def extract_from_summary(summary, fields, include_patterns=None, exclude_patterns=None):
-    """
-    Extract a fields from summary per test and histogram name
-    """
-    test_histo_value_map = {}
-    # need to re-arrange the JSON structure abit for per-test result pie charts
-    for histo_name, tests in summary.items():
-        # check if histo_name is in include patterns
-        if not check_patterns(histo_name, include_patterns, exclude_patterns):
-            continue
-        # loop over tests done
-        for test in tests:
-            test_name = test["test_name"]
-            if test_name not in test_histo_value_map:
-                test_histo_value_map[test_name] = {field: [] for field in fields}
-                test_histo_value_map[test_name]["histograms"] = []
-            if not test["comparable"]:
-                continue
-            test_histo_value_map[test_name]["histograms"].append(histo_name)
-            for field in fields:
-                test_histo_value_map[test_name][field].append(test[field])
-    return test_histo_value_map
-
-
-def plot_values_thresholds(summary, out_dir, title, include_patterns=None, exclude_patterns=None):
-    print("==> Plot values and thresholds <==")
-    test_histo_value_map = extract_from_summary(summary, ["value", "threshold"], include_patterns, exclude_patterns)
-
-    for which_test, histos_values_thresolds in test_histo_value_map.items():
-        if which_test == REL_VAL_TEST_SUMMARY_NAME:
-            continue
-        figure, ax = plt.subplots(figsize=(20, 20))
-        ax.plot(range(len(histos_values_thresolds["histograms"])), histos_values_thresolds["value"], label="values", marker="x")
-        ax.plot(range(len(histos_values_thresolds["histograms"])), histos_values_thresolds["threshold"], label="thresholds", marker="o")
-        ax.legend(loc="best", fontsize=20)
-        ax.set_xticks(range(len(histos_values_thresolds["histograms"])))
-        ax.set_xticklabels(histos_values_thresolds["histograms"], rotation=90)
-        ax.tick_params("both", labelsize=20)
-
-        figure.suptitle(f"{title} ({which_test})", fontsize=40)
-        save_path = join(out_dir, f"test_values_thresholds_{which_test}.png")
-        figure.tight_layout()
-        figure.savefig(save_path)
-        plt.close(figure)
 
 
 def plot_compare_summaries(summaries, fields, out_dir, *, labels=None, include_patterns=None, exclude_patterns=None):
@@ -377,7 +120,7 @@ def plot_compare_summaries(summaries, fields, out_dir, *, labels=None, include_p
         labels = [f"summary_{i}" for i, _ in enumerate(summaries)]
 
     for test_name in test_names:
-        if test_name == REL_VAL_TEST_SUMMARY_NAME:
+        if test_name == variables.REL_VAL_TEST_SUMMARY_NAME:
             continue
         histogram_names_intersection = []
         # First we figure out the intersection of histograms ==> histograms in common
@@ -412,268 +155,88 @@ def plot_compare_summaries(summaries, fields, out_dir, *, labels=None, include_p
         plt.close(figure)
 
 
-def plot_summary_grid(summary, flags, include_patterns, exclude_patterns, output_path):
-
-    print("==> Plot summary grid <==")
-
-    colors = [None] * len(REL_VAL_SEVERITY_MAP)
-    for name, color in REL_VAL_SEVERITY_COLOR_MAP.items():
-        colors[REL_VAL_SEVERITY_MAP[name]] = color
-    cmap = LinearSegmentedColormap.from_list("Custom", colors, len(colors))
-    collect_for_grid = []
-    collect_names = []
-    collect_annotations = []
-
-    for name, batch in summary.items():
-        if not check_patterns(name, include_patterns, exclude_patterns):
-            continue
-        include_this = not flags
-        collect_flags_per_test = [0] * len(REL_VAL_TEST_NAMES_MAP_SUMMARY)
-        collect_annotations_per_test = [""] * len(REL_VAL_TEST_NAMES_MAP_SUMMARY)
-        for test in batch:
-            test_name = test["test_name"]
-            if test_name not in REL_VAL_TEST_NAMES_MAP_SUMMARY:
-                continue
-            if flags and not include_this:
-                for f in flags:
-                    if test["result"] == f:
-                        include_this = True
-                        break
-
-            res = test["result"]
-            ind = REL_VAL_TEST_NAMES_MAP_SUMMARY[test_name]
-            if ind != len(REL_VAL_TEST_NAMES_MAP):
-                value_annotaion = f"{test['value']:.3f}" if test["comparable"] else "---"
-                collect_annotations_per_test[ind] = f"{test['threshold']:.3f}; {value_annotaion}"
-            collect_flags_per_test[ind] = REL_VAL_SEVERITY_MAP[res]
-
-        if not include_this:
-            continue
-        collect_for_grid.append(collect_flags_per_test)
-        collect_names.append(name)
-        collect_annotations.append(collect_annotations_per_test)
-
-    if not collect_for_grid:
-        print("WARNING: Nothing to plot for summary grid")
-        return
-
-    figure, ax = plt.subplots(figsize=(20, 20))
-    collect_for_grid = [c for _, c in sorted(zip(collect_names, collect_for_grid))]
-    collect_annotations = [c for _, c in sorted(zip(collect_names, collect_annotations))]
-    collect_names.sort()
-    seaborn.heatmap(collect_for_grid, ax=ax, cmap=cmap, vmin=-0.5, vmax=len(REL_VAL_SEVERITY_MAP) - 0.5, yticklabels=collect_names, xticklabels=REL_VAL_TEST_NAMES_SUMMARY, linewidths=0.5, annot=collect_annotations, fmt="")
-    cbar = ax.collections[0].colorbar
-    cbar.set_ticks(range(len(REL_VAL_SEVERITY_MAP)))
-    cbar.set_ticklabels(REL_VAL_SEVERITIES)
-    ax.set_title("Test summary (threshold; value)", fontsize=30)
-    figure.tight_layout()
-    figure.savefig(output_path)
-    plt.close(figure)
-
-
-def calc_thresholds(rel_val_dict, default_thresholds, margins_thresholds, args):
+def extract(input_filenames, target_filename, include_file_directories=None, add_if_exists=False, reference_extracted=None):
     """
-    calculate thresholds
+    Wrap the extraction of objects to be compared
+
+    Will be extracted (from TH1, QC objects, TTree etc.), converted to TH1 and put into a flat ROOT file structure.
+
+    Args:
+        reference_extracted: str
+        is used in case of the extraction of TTrees in which case the x-axis binning will be set according to that reference
+        to make objects comparable.
+
+        include_file_directories: list or None
+        will be passed to the ROOT macro and if not None, only sub-directories matching that will be browsed and extracted
     """
-    the_thresholds={}
+    if not include_file_directories:
+        include_file_directories = ""
 
-    if not args.use_values_as_thresholds:
-        for histo_name, tests in rel_val_dict.items():
-            this_histo_thresholds=[]
-            for t in tests:
-                test_name = t["test_name"]
-                if test_name == REL_VAL_TEST_SUMMARY_NAME:
-                    continue
-                these_thresholds={}
-                these_thresholds["test_name"] = test_name
-                these_thresholds["value"] = default_thresholds[test_name]
-                this_histo_thresholds.append(these_thresholds)
-            the_thresholds[histo_name] = this_histo_thresholds
+    if not reference_extracted:
+        reference_extracted = ""
 
+    if include_file_directories:
+        include_file_directories = ",".join(include_file_directories)
     else:
-        if args.use_values_as_thresholds[0].startswith("@"):
-            with open(args.use_values_as_thresholds[0][1:], "r") as f:
-                list_of_threshold_files = f.read().splitlines()
-        else:
-            list_of_threshold_files = args.use_values_as_thresholds
+        include_file_directories = ""
 
-        user_thresholds = []
-        for file_name in list_of_threshold_files:
-            with open(file_name, "r") as f:
-                user_thresholds.append(json.load(f).get("objects",{}))
+    if exists(target_filename) and not add_if_exists:
+        remove(target_filename)
 
-        for histo_name, tests in rel_val_dict.items():
-            this_histo_thresholds=[]
-            for t in tests:
-                these_thresholds={}
-                test_name = t["test_name"]
-                if test_name == REL_VAL_TEST_SUMMARY_NAME:
-                    continue
-                these_thresholds["test_name"] = test_name
-                threshold_list = []
-                for ut in user_thresholds:
-                    for ref_test in ut.get(histo_name, []):
-                        if ref_test["test_name"] == test_name:
-                            threshold_list.append(ref_test["value"])
-                if args.combine_thresholds == "mean":
-                    tuned_threshold = pow(margins_thresholds[test_name],REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[test_name]]) * sum(threshold_list) / len(threshold_list)
-                else:
-                    if REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[test_name]] == 1:
-                        maxmin = max(threshold_list)
-                    else:
-                        maxmin = min(threshold_list)
-                    tuned_threshold = pow(margins_thresholds[test_name],REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[test_name]]) * maxmin
-                if args.combine_tuned_and_fixed_thresholds:
-                    if REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[test_name]] == 1:
-                        these_thresholds["value"] = max(tuned_threshold,default_thresholds[test_name])
-                    else:
-                        these_thresholds["value"] = min(tuned_threshold,default_thresholds[test_name])
-                else:
-                    these_thresholds["value"] = tuned_threshold
-                this_histo_thresholds.append(these_thresholds)
-            the_thresholds[histo_name] = this_histo_thresholds
+    # The ROOT macro is run inside the cwd and puts basename there
+    cwd = dirname(target_filename)
+    target_filename = basename(target_filename)
+    log_file_name = f"{target_filename}_extract_and_flatten.log"
 
-    return the_thresholds
+    print(f"Extraction of files\n{','.join(input_filenames)}")
+
+    for f in input_filenames:
+        f = abspath(f)
+        cmd = f"\\(\\\"{f}\\\",\\\"{target_filename}\\\",\\\"{reference_extracted}\\\",\\\"{include_file_directories}\\\"\\)"
+        cmd = f"root -l -b -q {ROOT_MACRO_EXTRACT}{cmd}"
+        run_macro(cmd, log_file_name, cwd)
 
 
-def write_single_summary(comp_objects, meta_info, path):
-    with open(path, "w") as f:
-        json.dump({"objects": comp_objects, "meta_info": meta_info}, f, indent=2)
-
-
-def read_single_summary(path):
-    with open(path, "r") as f:
-        d = json.load(f)
-        return d.get("objects", {}), d.get("meta_info", {})
-
-
-def make_single_meta_info(args):
-    return {"batch_i": [abspath(path) for path in args.input1], "batch_j": [abspath(path) for path in args.input2]}
-
-
-def make_single_summary(rel_val_dict, args, output_dir, include_patterns=None, exclude_patterns=None, flags=None, flags_summary=None):
-    """
-    Make the usual summary
-    """
-    def assign_result_flag(is_critical, comparable, passed):
-        result = "GOOD"
-        if is_critical:
-            if not comparable:
-                result = "CRIT_NC"
-            elif not passed:
-                result = "BAD"
-        else:
-            if not comparable:
-                result = "NONCRIT_NC"
-            elif not passed:
-                result = "WARNING"
-        return result
-
-    this_summary = {}
-
-    default_thresholds = {t: getattr(args, f"{t}_threshold") for t in REL_VAL_TEST_NAMES}
-    margins_thresholds = {t: getattr(args, f"{t}_threshold_margin") for t in REL_VAL_TEST_NAMES}
-    test_enabled = {t: getattr(args, f"with_{t}") for t in REL_VAL_TEST_NAMES}
-    if not any(test_enabled.values()):
-        test_enabled = {t: True for t in test_enabled}
-
-    the_thresholds = calc_thresholds(rel_val_dict, default_thresholds, margins_thresholds, args)
-    with open(join(output_dir, "used_thresholds.json"), "w") as f:
-            json.dump(the_thresholds, f, indent=2)
-
-    for histo_name, tests in rel_val_dict.items():
-        if not check_patterns(histo_name, include_patterns, exclude_patterns):
-            continue
-        test_summary = {"test_name": REL_VAL_TEST_SUMMARY_NAME,
-                        "value": None,
-                        "threshold": None,
-                        "result": None}
-        is_critical_summary = False
-        passed_summary = True
-        is_comparable_summary = True
-        these_tests = []
-        for t in tests:
-            if t["test_name"] == REL_VAL_TEST_SUMMARY_NAME or not test_enabled[t["test_name"]]:
-                continue
-            test_name = t["test_name"]
-            test_id = REL_VAL_TEST_NAMES_MAP[test_name]
-            threshold = the_thresholds[histo_name][REL_VAL_TEST_NAMES_MAP[test_name]]["value"]
-            t["threshold"] = threshold
-
-            comparable = t["comparable"]
-            passed = True
-            is_critical = REL_VAL_TEST_CRITICAL[test_id] or histo_name.find("_ratioFromTEfficiency") != -1
-            if comparable:
-                passed = t["value"]*REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[t["test_name"]]] <=  threshold*REL_VAL_TEST_UPPER_LOWER_THRESHOLD[REL_VAL_TEST_NAMES_MAP[t["test_name"]]]
-
-            t["result"] = assign_result_flag(is_critical, comparable, passed)
-
-            is_critical_summary = is_critical or is_critical_summary
-            is_comparable_summary = comparable and is_comparable_summary
-            if is_critical:
-                # only mark potentially as failed if we run over a critical test
-                passed_summary = passed_summary and passed
-            these_tests.append(t)
-
-        test_summary["result"] = assign_result_flag(is_critical_summary, is_comparable_summary, passed_summary)
-        test_summary["comparable"] = is_comparable_summary
-        these_tests.append(test_summary)
-        if not check_flags(these_tests, flags, flags_summary):
-            continue
-        this_summary[histo_name] = these_tests
-
-    return this_summary
-
-
-def rel_val_files(files1, files2, args, output_dir):
+def rel_val_files(files1, files2, args, output_dir, no_extract=False):
     """
     RelVal for 2 ROOT files, simply a wrapper around ReleaseValidation.C macro
+
+    Args:
+        no_extract: bool
+        If True, expect files1 and files2 to be of length 1. These are expected to already contain extracted objects.
+        Hence, extraction will be skipped and objects from these files will directly be compared.
+        Potential previous results will not be overwritten but the new ones will be dumped into a new directory (as usual)
     """
-    def run_macro(cmd, log_file):
-        p = Popen(split(cmd), cwd=output_dir, stdout=PIPE, stderr=STDOUT, universal_newlines=True)
-        log_file = open(log_file, 'a')
-        for line in p.stdout:
-            log_file.write(line)
-        p.wait()
-        log_file.close()
 
     print("==> Process and compare 2 sets of files <==")
 
+    # prepare the output directory
     if not exists(output_dir):
         makedirs(output_dir)
-    log_file_extract = join(abspath(output_dir), "extract_and_flatten.log")
     log_file_rel_val = join(abspath(output_dir), "rel_val.log")
-    if args.include_dirs:
-        include_directories = ",".join(args.include_dirs)
+
+    # flat ROOT files to extract to and read from during RelVal; make absolute paths so we don't confuse ourselves when running e.g. ROOT macros in different directories
+    file_1 = abspath(join(output_dir, "extracted_objects_1.root"))
+    file_2 = abspath(join(output_dir, "extracted_objects_2.root"))
+
+    if no_extract:
+        # in this case we expect the input files to be what we would otherwise extract firt
+        if len(files1) != 1 or len(files2) != 1:
+            print(f"ERROR: --no-extract option was passed and expecting list of files to be of length 1 each. However, received lengths of {len(files1)} and {len(files2)}")
+            return 1
+        file_1 = abspath(files1[0])
+        file_2 = abspath(files2[0])
     else:
-        include_directories = ""
+        # extract all objects and put into flat ROOT file structure
+        extract(files1, file_1, args.include_dirs, args.add)
+        extract(files2, file_2, args.include_dirs, args.add, reference_extracted=file_1)
 
-    print(f"Extraction of files\n{','.join(files1)}")
-    file_1 = "newfile1.root"
-    file_from_here = join(output_dir, file_1)
-    if exists(file_from_here) and not args.add:
-        remove(file_from_here)
-    for f in files1:
-        f = abspath(f)
-        cmd = f"\\(\\\"{f}\\\",\\\"{file_1}\\\",\\\"\\\",\\\"{include_directories}\\\"\\)"
-        cmd = f"root -l -b -q {ROOT_MACRO_EXTRACT}{cmd}"
-        run_macro(cmd, log_file_extract)
-
-    print(f"Extraction of files\n{','.join(files2)}")
-    file_2 = "newfile2.root"
-    file_from_here = join(output_dir, file_2)
-    if exists(file_from_here) and not args.add:
-        remove(file_from_here)
-    for f in files2:
-        f = abspath(f)
-        cmd = f"\\(\\\"{f}\\\",\\\"{file_2}\\\",\\\"{file_1}\\\",\\\"{include_directories}\\\"\\)"
-        cmd = f"root -l -b -q {ROOT_MACRO_EXTRACT}{cmd}"
-        run_macro(cmd, log_file_extract)
-
+    # RelVal on flattened files
     cmd = f"\\(\\\"{file_1}\\\",\\\"{file_2}\\\",{args.test},\\\"{args.labels[0]}\\\",\\\"{args.labels[1]}\\\"\\)"
     cmd = f"root -l -b -q {ROOT_MACRO_RELVAL}{cmd}"
     print("Running RelVal on extracted objects")
-    run_macro(cmd, log_file_rel_val)
+    run_macro(cmd, log_file_rel_val, cwd=output_dir)
+
     # This comes from the ROOT macro
     json_path = join(output_dir, "RelVal.json")
 
@@ -691,84 +254,12 @@ def rel_val_files(files1, files2, args, output_dir):
     final_summary = make_single_summary(rel_val_summary, args, output_dir)
     meta_info = make_single_meta_info(args)
     write_single_summary(final_summary, meta_info, join(output_dir, "Summary.json"))
-
     plot_pie_charts(final_summary, output_dir, "")
     plot_values_thresholds(final_summary, output_dir, "")
     plot_summary_grid(final_summary, None, None, None, join(output_dir, "SummaryTests.png"))
+    print_summary(final_summary, long=args.long)
 
     return 0
-
-
-def map_histos_to_severity(summary, include_patterns=None, exclude_patterns=None):
-    """
-    Map the histogram names to their severity of the test
-    """
-    test_n_hist_map = {s: [] for i, s in enumerate(REL_VAL_SEVERITIES) if REL_VAL_SEVERITIES_USE_SUMMARY[i]}
-
-    # need to re-arrange the JSON structure abit for per-test result pie charts
-    for histo_name, tests in summary.items():
-        # check if histo_name is in include_patterns
-        if not check_patterns(histo_name, include_patterns, exclude_patterns):
-            continue
-        # loop over tests done
-        for test in tests:
-            test_name = test["test_name"]
-            if test_name != REL_VAL_TEST_SUMMARY_NAME:
-                continue
-            result = test["result"]
-            test_n_hist_map[result].append(histo_name)
-
-    return test_n_hist_map
-
-
-def print_summary(summary, include_patterns=None, exclude_patterns=None, long=False):
-    """
-    Check if any 2 histograms have a given severity level after RelVal
-    """
-
-    test_n_hist_map = map_histos_to_severity(summary, include_patterns, exclude_patterns)
-
-    n_all = sum(len(v) for v in test_n_hist_map.values())
-    print(f"\n#####\nNumber of compared histograms: {n_all}\nBased on critical tests, severities are\n")
-    for sev, histos in test_n_hist_map.items():
-        print(f"  {sev}: {len(histos)}")
-        if long:
-            for i, h in enumerate(histos, start=1):
-                print(f"    {i}. {h}")
-    print("#####\n")
-
-
-def make_global_summary(in_dir):
-    """
-    Make a summary per histogram (that should be able to be parsed by Grafana eventually)
-    """
-    file_paths = glob(f"{in_dir}/**/Summary.json", recursive=True)
-    summary = {}
-
-    batch_i = []
-    batch_j = []
-
-    for path in file_paths:
-        # go through all we found
-        current_summary, meta_info = read_single_summary(path)
-        batch_i.extend(meta_info.get("batch_i", []))
-        batch_j.extend(meta_info.get("batch_j", []))
-
-        # remove the file name, used as the top key for this collection
-        rel_val_path = "/".join(path.split("/")[:-1])
-        type_specific = relpath(rel_val_path, in_dir)
-        rel_path_plot = join(type_specific, "overlayPlots")
-        type_global = type_specific.split("/")[0]
-        for histo_name, tests in current_summary.items():
-            summary[histo_name] = tests
-            # loop over tests done
-            for test in tests:
-                test["name"] = histo_name
-                test["type_global"] = type_global
-                test["type_specific"] = type_specific
-                test["rel_path_plot"] = join(rel_path_plot, f"{histo_name}.png")
-
-    return summary, {"batch_i": batch_i, "batch_j": batch_j}
 
 
 def rel_val(args):
@@ -777,11 +268,10 @@ def rel_val(args):
     """
     if args.add:
         print(f"NOTE: Extracted objects will be added to existing ones in case there was already a RelVal at {args.output}.\n")
-    func = None
     # construct the bit mask
     args.test = 0
     default_sum = 0
-    for i, t in enumerate(REL_VAL_TEST_NAMES):
+    for i, t in enumerate(variables.REL_VAL_TEST_NAMES):
         bit = 2**i
         args.test += bit * getattr(args, f"with_{t}")
         default_sum += bit
@@ -789,22 +279,8 @@ def rel_val(args):
         args.test = default_sum
     if not exists(args.output):
         makedirs(args.output)
-    rel_val_files(args.input1, args.input2, args, args.output)
-    global_summary, meta_info = make_global_summary(args.output)
-    write_single_summary(global_summary, meta_info, join(args.output, "SummaryGlobal.json"))
-    print_summary(global_summary, long=args.long)
+    rel_val_files(args.input1, args.input2, args, args.output, args.no_extract)
     return 0
-
-
-def get_filepath(d):
-    summary_global = join(d, "SummaryGlobal.json")
-    if exists(summary_global):
-        return summary_global
-    summary = join(d, "Summary.json")
-    if exists(summary):
-        return summary
-    print(f"Can neither find {summary_global} nor {summary}. Nothing to work with.")
-    return None
 
 
 def copy_overlays(path, output_dir,summary):
@@ -828,12 +304,9 @@ def inspect(args):
     """
     Inspect a Summary.json in view of RelVal severity
     """
-    path = args.path
-
-    if isdir(path):
-        path = get_filepath(path)
-        if not path:
-            return 1
+    path = get_summary_path(args.path)
+    if not path:
+        return 1
 
     output_dir = args.output or join(dirname(path), "user_summary")
     if not exists(output_dir):
@@ -871,13 +344,14 @@ def compare(args):
 
     # load everything
     include_patterns, exclude_patterns = load_patterns(args.include_patterns, args.exclude_patterns)
-    summaries = find_mutual_files(inputs, "SummaryGlobal.json")
-    if not summaries:
-        print(f"Cannot find \"SummaryGlobal.json\" in given directories {inputs[0]} and {inputs[1]}. Do the directories exist?")
+    inputs = [join(get_summary_path(input_path)) for input_path in inputs]
+
+    if not all(inputs):
+        print(f"ERROR: Cannot find {inputs[0]} and {inputs[1]}")
         return 1
-    summaries = [join(i, s) for i in inputs for s in summaries]
-    for i, _ in enumerate(summaries):
-        summaries[i], _ = read_single_summary(summaries[i])
+
+    # only read the summaries, without meta info
+    summaries = [read_single_summary(input_path)[0] for input_path in inputs]
 
     if not args.difference and not args.compare_values:
         args.difference, args.compare_values = (True, True)
@@ -893,7 +367,7 @@ def compare(args):
         s = "\nCOMPARING RELVAL SUMMARY\n"
         summaries = [map_histos_to_severity(summary, include_patterns, exclude_patterns) for summary in summaries]
         print("Histograms with different RelVal results from 2 RelVal runs")
-        for severity, use in zip(REL_VAL_SEVERITY_MAP, REL_VAL_SEVERITIES_USE_SUMMARY):
+        for severity, use in zip(variables.REL_VAL_SEVERITY_MAP, variables.REL_VAL_SEVERITIES_USE_SUMMARY):
             if not use:
                 continue
             intersection = list(set(summaries[0][severity]) & set(summaries[1][severity]))
@@ -919,7 +393,7 @@ def influx(args):
     Create an influxDB metrics file
     """
     output_dir = args.dir
-    json_in = join(output_dir, "SummaryGlobal.json")
+    json_in = join(output_dir, "Summary.json")
     if not exists(json_in):
         print(f"Cannot find expected JSON summary {json_in}.")
         return 1
@@ -946,12 +420,12 @@ def influx(args):
         for i, (histo_name, tests) in enumerate(summary.items()):
             if not tests:
                 continue
-            common_string = f"{row_tags},type_global={tests[0]['type_global']},type_specific={tests[0]['type_specific']},id={i}"
+            common_string = f"{row_tags},id={i}"
             if args.web_storage:
                 common_string += f",web_storage={join(args.web_storage, tests[0]['rel_path_plot'])}"
             common_string += f",histogram_name={histo_name}"
             for test in tests:
-                test_string = common_string + f",test_name={test['test_name']} status={REL_VAL_SEVERITY_MAP[test['result']]}"
+                test_string = common_string + f",test_name={test['test_name']} status={variables.REL_VAL_SEVERITY_MAP[test['result']]}"
                 for key in ("value", "threshold"):
                     value = test[key]
                     if value is None:
@@ -983,11 +457,9 @@ def print_table(args):
     """
     Print the filtered histogram names of a Summary.json as list to screen
     """
-    path = args.path
-    if isdir(path):
-        path = get_filepath(path)
-        if not path:
-            return 1
+    path = get_summary_path(args.path)
+    if not path:
+        return 1
 
     include_patterns, exclude_patterns = load_patterns(args.include_patterns, args.exclude_patterns, False)
     summary, _ = read_single_summary(path)
@@ -1013,12 +485,13 @@ def main():
     common_file_parser.add_argument("-i", "--input1", nargs="*", help="EITHER first set of input files for comparison OR first input directory from simulation for comparison", required=True)
     common_file_parser.add_argument("-j", "--input2", nargs="*", help="EITHER second set of input files for comparison OR second input directory from simulation for comparison", required=True)
     common_file_parser.add_argument("--labels", nargs=2, help="labels you want to appear in the plot legends in case of overlay plots from batches -i and -j", default=("batch_i", "batch_j"))
+    common_file_parser.add_argument("--no-extract", dest="no_extract", action="store_true", help="no extraction but immediately expect histograms present for comparison")
 
     common_threshold_parser = argparse.ArgumentParser(add_help=False)
     common_threshold_parser.add_argument("--use-values-as-thresholds", nargs="*", dest="use_values_as_thresholds", help="Use values from another run as thresholds for this one")
     common_threshold_parser.add_argument("--combine-thresholds", dest="combine_thresholds",  choices=["mean", "max/min"], help="Arithmetic mean or maximum/minimum is chosen as threshold value", default="mean")
     common_threshold_parser.add_argument("--combine-tuned-and-fixed-thresholds",dest="combine_tuned_and_fixed_thresholds", action="store_true", help="Combine the result from 'combine-thresholds' with the fixed threshold value using maximum/minimum")
-    for test, thresh in zip(REL_VAL_TEST_NAMES, REL_VAL_TEST_DEFAULT_THRESHOLDS):
+    for test, thresh in zip(variables.REL_VAL_TEST_NAMES, variables.REL_VAL_TEST_DEFAULT_THRESHOLDS):
         test_dashed = test.replace("_", "-")
         common_threshold_parser.add_argument(f"--with-test-{test_dashed}", dest=f"with_{test}", action="store_true", help=f"run {test} test")
         common_threshold_parser.add_argument(f"--test-{test_dashed}-threshold", dest=f"{test}_threshold", type=float, help=f"{test} threshold", default=thresh)
@@ -1030,8 +503,8 @@ def main():
     common_pattern_parser.add_argument("--exclude-patterns", dest="exclude_patterns", nargs="*", help="exclude objects whose name includes at least one of the given patterns")
 
     common_flags_parser = argparse.ArgumentParser(add_help=False)
-    common_flags_parser.add_argument("--flags", nargs="*", help="extract all objects which have at least one test with this severity flag", choices=list(REL_VAL_SEVERITY_MAP.keys()))
-    common_flags_parser.add_argument("--flags-summary", dest="flags_summary", nargs="*", help="extract all objects which have this severity flag as overall test result", choices=list(REL_VAL_SEVERITY_MAP.keys()))
+    common_flags_parser.add_argument("--flags", nargs="*", help="extract all objects which have at least one test with this severity flag", choices=list(variables.REL_VAL_SEVERITY_MAP.keys()))
+    common_flags_parser.add_argument("--flags-summary", dest="flags_summary", nargs="*", help="extract all objects which have this severity flag as overall test result", choices=list(variables.REL_VAL_SEVERITY_MAP.keys()))
 
     common_verbosity_parser = argparse.ArgumentParser(add_help=False)
     common_verbosity_parser.add_argument("--long", action="store_true", help="enhance verbosity")
@@ -1044,7 +517,7 @@ def main():
     rel_val_parser.set_defaults(func=rel_val)
 
     inspect_parser = sub_parsers.add_parser("inspect", parents=[common_threshold_parser, common_pattern_parser, common_flags_parser, common_verbosity_parser])
-    inspect_parser.add_argument("path", help="either complete file path to a Summary.json or SummaryGlobal.json or directory where one of the former is expected to be")
+    inspect_parser.add_argument("path", help="either complete file path to a Summary.json or directory where one of the former is expected to be")
     inspect_parser.add_argument("--plot", action="store_true", help="Plot the summary grid")
     inspect_parser.add_argument("--output", "-o", help="output directory, by default points to directory where the Summary.json was found")
     inspect_parser.add_argument("--copy-overlays", dest="copy_overlays", action="store_true", help="Copy overlay plots that meet the filter criteria to output directory")
@@ -1064,7 +537,7 @@ def main():
     influx_parser.set_defaults(func=influx)
 
     print_parser = sub_parsers.add_parser("print", parents=[common_pattern_parser, common_flags_parser])
-    print_parser.add_argument("path", help="either complete file path to a Summary.json or SummaryGlobal.json or directory where one of the former is expected to be")
+    print_parser.add_argument("path", help="either complete file path to a Summary.json or directory where one of the former is expected to be")
     print_parser.set_defaults(func=print_table)
 
     file_size_parser = sub_parsers.add_parser("file-sizes", parents=[common_file_parser])
